* lots-o-files: applied some of the changes from Christoph Reichenbach <jameson@linuxgames.com>   I didn't incorporate all changes yet, though. * exult.cc: on mac, META (=Command key) is now used instead of ALT

+++ objs/spellbook.cc
@@ -40,7 +40,7 @@ Spellbook_object::Spellbook_object
 	unsigned int shapex, unsigned int shapey, 
 	unsigned int lft, 
 	unsigned char *c,		// Circle spell flags.
-	unsigned long f			// Flags (unknown).
+	uint32 f			// Flags (unknown).
 	) : Ireg_game_object(shapenum, framenum,
 			shapex, shapey, lft), flags(f), bookmark(-1)
 	{
